@@ -11,18 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
+    "@angular/forms": "^8.2.14",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "angular7-csv": "^0.2.12",
@@ -34,21 +33,21 @@
     "lodash": "^4.17.11",
     "ngx-cookie-service": "^2.1.0",
     "ngx-gauge": "^1.0.0-beta.7",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.4",
     "tslib": "^1.9.0",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.29"
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "^7.2.13",
+    "@angular-devkit/build-angular": "~0.803.25",
+    "@angular/cli": "~8.3.25",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.123",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -59,6 +58,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
