@@ -14,7 +14,7 @@ import { MessageService } from '../../services/message.service';
 export class LoginComponent implements OnInit {
 
   // creates a reference to the first input element
-  @ViewChild('usernameEle') private usernameElement: ElementRef;
+  @ViewChild('usernameEle', { static: true }) private usernameElement: ElementRef;
 
   // shows and hides the password
   hide: boolean;
